requirements.txt
streamlit
transformers
torch
gTTS
opencv-python
deep-translator
pillow
install 